Fix bug, now only add includes to tests instead of all targets
Fixes bug that modules have access to include directories of modules
they have not added as dependencies.
Previous behaviour was that if tests existed in a module,
the "test include-paths" were added with "INCLUDE_DIRECTORIES"
which adds them to general includes paths and this way to all targets.
By using set_target_properties, the test include directories are only
added to the test targets.

Change-Id: I291059bd5a91a04de90102b0d257f4d538a6a160